Player properly moves between levels and more
Added constructor and method to allow the randomization of coordinates
between minimum x y values and maximum x y values, also added the
ability to properly move between levels and appear at the appropriate
door in the next level. Added the functionality to spawn multiple
different kinds of enemies and items with the passing of the available
kinds in an array of characters. Included a couple of new exceptions.